requirements.txt
opencv-python==4.9.0.80
pillow==10.2.0
PyAutoGUI==0.9.54